+++ lua/haskell-tools/repl.lua
@@ -77,6 +77,10 @@ function repl.mk_repl_cmd(file)
       return nil
     end
   end
+  local opts = ht.config.options.tools.repl
+  if opts.prefer == 'stack' and project.is_stack_project(chk_path) then
+    return mk_stack_repl_cmd(file)
+  end
   if project.is_cabal_project(chk_path) then
     return mk_cabal_repl_cmd(file)
   end