+++ tests/repl_spec.lua
@@ -30,3 +30,29 @@ if deps.has_toggleterm() then
   mk_repl_setup_test('toggleterm', true)
   mk_repl_setup_test('toggleterm', false)
 end
+
+local repl = assert(ht.repl, 'repl not set up')
+local cwd = vim.fn.getcwd()
+
+describe('mk_repl_cmd', function()
+  local stack_project_file = cwd .. '/tests/fixtures/stack/single-package/Abc.hs'
+  if vim.fn.executable('stack') == 1 then
+    it('prefers stack if stack.yml exists', function()
+      local cmd = assert(repl.mk_repl_cmd(stack_project_file))
+      assert(#cmd > 1)
+      assert.same('stack', cmd[1])
+    end)
+  else
+    it('prefers cabal even if stack.yml exists if cabal files exist', function()
+      local cmd = assert(repl.mk_repl_cmd(stack_project_file))
+      assert(#cmd > 1)
+      assert.same('cabal', cmd[1])
+    end)
+  end
+  local cabal_project_file = cwd .. '/tests/fixtures/cabal/single-package/Abc.hs'
+  it('prefers cabal if no stack.yml exists and cabal files exist', function()
+    local cmd = assert(repl.mk_repl_cmd(cabal_project_file))
+    assert(#cmd > 1)
+    assert.same('cabal', cmd[1])
+  end)
+end)