+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.11.0] - 2023-07-05
 ### Changed
 - Improvements to type signature detection from `textDocument/hover` docs.
 
 ### Added
 - Hover: Hoogle search entries for all detected type signatures.
+
+### Fixed
+- repl: If both stack and cabal files are presend, prefer stack if it is installed.
+  This is configurable with the option `tools.repl.prefer`.
 
 ## [1.10.2] - 2023-05-22
 ### Fixed